Wiper-hazard fix.

+++ include/publish_control.h
@@ -44,7 +44,7 @@ class PublishControl
     static bool brake_0_rcvd;
     static int headlight_state;
     static bool headlight_state_change;
-    static int wiper_state;
+    static uint16_t wiper_state;
 
   protected:
     virtual void check_is_enabled(const sensor_msgs::Joy::ConstPtr& msg);
